Separate `UserRepository` -> `withJWT`, `withoutJWT`
