+++ app/src/main/java/kr/khs/oneboard/repository/UserRepositoryImpl.kt
@@ -6,39 +6,12 @@ import kr.khs.oneboard.api.ApiService
 import kr.khs.oneboard.data.Lecture
 import kr.khs.oneboard.data.api.Response
 import kr.khs.oneboard.utils.SUCCESS
+import javax.inject.Inject
+import javax.inject.Named
 
-class UserRepositoryImpl(val apiService: ApiService) : UserRepository {
-    override suspend fun healthCheck(): Boolean {
-        val response: Response<Boolean>
-        withContext(Dispatchers.IO) {
-//            response = apiService.healthCheck()
-            response = Response(SUCCESS, true)
-        }
-        return response.data
-    }
-
-    override suspend fun login(email: String, password: String): Boolean {
-        // TODO: 2021/10/20 로그인 성공 시 토큰 저장
-        val response: Response<Boolean>
-        withContext(Dispatchers.IO) {
-//            response = apiService.login()
-            response = Response(SUCCESS, true)
-        }
-        return response.data
-    }
-
-    override suspend fun loginCheck(token: String): Boolean {
-        val response: Response<Boolean>
-        withContext(Dispatchers.IO) {
-//            response = apiService.loginCheck()
-            response = Response(SUCCESS, false)
-        }
-        return response.data
-    }
-
-    override suspend fun logout() {
-        TODO("Not yet implemented")
-    }
+class UserRepositoryImpl @Inject constructor(
+    @Named("withJWT") private val apiService: ApiService,
+) : UserRepository {
 
     override suspend fun getUserInfo() {
         TODO("Not yet implemented")
